requirements.txt
influxdb-client
python-dotenv
pandas